requirements.txt
torch>=2.1.0
torchvision>=0.16.0
timm==0.9.12
einops==0.7.0
scikit-learn==1.4.0
pandas==2.2.0
matplotlib==3.8.2
seaborn==0.13.1
opencv-python==4.9.0.80
albumentations==1.3.1
pyyaml==6.0.1
tqdm==4.66.1
tensorboard==2.15.1
numpy>=1.24.0
Pillow>=10.0.0
